Fixing division by int issue

+++ CUDA/vecAdd.cu
@@ -72,7 +72,7 @@ int main( int argc, char* argv[] )
     double sum = 0;
     for(i=0; i<n; i++)
         sum += h_c[i];
-    printf("final result: %f\n", sum/n);
+    printf("final result: %f\n", sum/(double)n);
  
     // Release device memory
     cudaFree(d_a);
