+++ OpenACC/vecAdd.c
@@ -40,7 +40,7 @@ int main( int argc, char* argv[] )
     for(i=0; i<n; i++) {
         sum += c[i];
     }
-    sum = sum/n;
+    sum = sum/(double)n;
     printf("final result: %f\n", sum);
  
     // Release memory
